translate/requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-openai
langsmith
openai
transformers
faiss-cpu
confluent-kafka
boto3
python-dotenv
typing_extensions>=4.8.0
javalang